requirements.txt
tensorflow-gpu==2.1.0
matplotlib==3.1.3
Pillow==6.2.0
